% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{gems_version}
\alias{gems_version}
\title{Versao do pacote GEMSFertilizer}
\usage{
gems_version()
}
\value{
String com a versao atual do pacote.
}
\description{
Versao do pacote GEMSFertilizer
}
\examples{
GEMSFertilizer::gems_version()
}
