% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\docType{package}
\name{GEMSFertilizer-package}
\alias{GEMSFertilizer}
\alias{GEMSFertilizer-package}
\title{GEMSFertilizer: Sistema de Recomendacao de Adubacao e Calagem para Culturas Agricolas}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Aplicativo Shiny interativo para recomendacao de adubacao mineral (N, P2O5, K2O), calagem e gessagem de culturas agricolas. Baseado no Manual de Adubacao e Calagem para Minas Gerais - 5a Aproximacao (2023) e no Manual de Recomendacoes de Adubacao e Calagem do Estado de Sergipe (EMBRAPA). Inclui modulo de pesquisa agricola com conversao de doses por hectare para unidades experimentais (metro linear, cova, vaso e parcela), busca de precos de fertilizantes via CEPEA/AgroLink, e graficos interativos de fertilidade do solo. Culturas suportadas: milho, feijao, cana-de-acucar, arroz, mandioca, amendoim, sorgo, pastagem e abacaxi.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seu_usuario/GEMSFertilizer}
  \item Report bugs at \url{https://github.com/seu_usuario/GEMSFertilizer/issues}
}

}
\author{
\strong{Maintainer}: GEMS Research Group \email{gustavooliveira@academico.ufs.br} (\href{https://orcid.org/0000-0002-3839-6261}{ORCID})

Authors:
\itemize{
  \item GEMS Research Group \email{gustavooliveira@academico.ufs.br} (\href{https://orcid.org/0000-0002-3839-6261}{ORCID})
}

}
\keyword{internal}
